more AndroidTV changes

+++ android/app/src/main/java/com/izzy/kart/TVActivity.java
@@ -1,6 +1,7 @@
 package com.izzy.kart;
 
 import android.view.KeyEvent;
+import android.view.InputDevice;
 
 public class TVActivity extends MainActivity{
     @Override
@@ -8,17 +9,24 @@ public class TVActivity extends MainActivity{
         return true;
     }
 
-    @Override
-    public boolean dispatchKeyEvent(KeyEvent event) {
-        if (event.getKeyCode() == KeyEvent.KEYCODE_BACK) {
-            if (event.getAction() == KeyEvent.ACTION_UP) {
-                // Finish the activity
-                finish();
-                // Kill the app process explicitly to fully terminate the app
-                android.os.Process.killProcess(android.os.Process.myPid());
-            }
-            return true;
-        }
-        return super.dispatchKeyEvent(event);
-    }
+	@Override
+	public boolean dispatchKeyEvent(KeyEvent event) {
+		if (event.getKeyCode() == KeyEvent.KEYCODE_BACK) {
+			int source = event.getSource();
+
+			boolean isKeyboard = (source & InputDevice.SOURCE_KEYBOARD) == InputDevice.SOURCE_KEYBOARD;
+			boolean isDpad = (source & InputDevice.SOURCE_DPAD) == InputDevice.SOURCE_DPAD;
+			boolean isGamepad = (source & InputDevice.SOURCE_GAMEPAD) == InputDevice.SOURCE_GAMEPAD;
+
+			if (isKeyboard && !isDpad && !isGamepad) {
+				if (event.getAction() == KeyEvent.ACTION_UP) {
+					finish();
+					android.os.Process.killProcess(android.os.Process.myPid());
+				}
+				return true;
+			}
+			return false;
+		}
+		return super.dispatchKeyEvent(event);
+	}
 }
